+++ CHANGELOG.md
@@ -5,10 +5,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support for parsing message links to [yuyo.links][].
+
 ### Changed
 - Added support for `ptb.` and `canary.` links to [yuyo.links][].
 - Modals now default the per-field custom IDs (i.e. for text components) to the parameter's
   name (attribute name for [ModalOptions][yuyo.modals.ModalOptions] fields), if set.
+- Renamed the link get and fetch methods:
+    * `InviteLink.fetch` to [InviteLink.fetch_invite][yuyo.links.InviteLink.fetch_invite]
+    * `InviteLink.get` to [InviteLink.get_invite][yuyo.links.InviteLink.get_invite]
+    * `MessageLink.fetch` to [MessageLink.fetch_message][yuyo.links.MessageLink.fetch_message]
+    * `MessageLink.get` to [MessageLink.get_message][yuyo.links.MessageLink.get_message]
+    * `TemplateLink.fetch` to [TemplateLink.fetch_template][yuyo.links.TemplateLink.fetch_template]
+    * `WebhookLink.fetch` to [WebhookLink.fetch_webhook][yuyo.links.WebhookLink.fetch_webhook]
 
 ### Fixed
 - [ModalOptions][yuyo.modals.ModalOptions] attributes now correctly expose the values passed to